refactor: Separate ProductDetailViewModel from ProductViewModel (#40)
- Created ProductDetailViewModel to manage product detail specific logic.
- Moved product detail related state and functions from ProductViewModel to ProductDetailViewModel.
- Updated MainView and ProductDetailScreen to use the new ProductDetailViewModel.
- Adjusted ProductDetailScreen to fetch product details and request list based on productId passed as a parameter.
- Removed unused product ID setting logic from MainView.
